requirements.txt
streamlit
boto3
langchain
langchain-aws
langgraph
pandas
psycopg2-binary
python-dotenv
